+++ src/tests/Package.test.ts
@@ -20,6 +20,7 @@ import { IRepository } from "../Management/IRepository.js";
 import { PackagePerson } from "../Management/PackagePerson.js";
 import { Person } from "../Management/Person.js";
 import { Package } from "../Package.js";
+import { PackageType } from "../PackageType.js";
 import { JSONObject } from "../Utilities/JSONObject.js";
 import { PropertyChecker } from "./PropertyChecker.js";
 import { TestContext } from "./TestContext.js";
@@ -199,6 +200,7 @@ export function PackageTests(context: TestContext): void
                     metadata = {
                         name: text(),
                         version: `${digit()}.${digit()}.${digit()}`,
+                        type: context.Random.bool() ? PackageType.CommonJS : PackageType.ESModule,
                         private: context.Random.bool(),
                         description: text(100),
                         author: person(),
@@ -229,6 +231,21 @@ export function PackageTests(context: TestContext): void
                         cpu: [
                             arch()
                         ],
+                        exports: {
+                            ".": {
+                                import: text(),
+                                require: text(),
+                                types: text(),
+                                default: text()
+                            }
+                        },
+                        imports: {
+                            "#dep": {
+                                import: text(),
+                                require: text(),
+                                default: text()
+                            }
+                        },
                         main: "./lib/index.js",
                         types: "./lib/index.d.ts",
                         browser: text(),
@@ -292,6 +309,7 @@ export function PackageTests(context: TestContext): void
             {
                 strictEqual(npmPackage.Name, metadata.name);
                 strictEqual(npmPackage.Version, metadata.version);
+                strictEqual(npmPackage.Type, metadata.type);
                 strictEqual(npmPackage.Private, metadata.private);
                 strictEqual(npmPackage.Description, metadata.description);
                 AssertPerson(npmPackage.Author, metadata.author);
@@ -302,6 +320,8 @@ export function PackageTests(context: TestContext): void
                 AssertDictionary(npmPackage.Engines, metadata.engines);
                 deepStrictEqual(npmPackage.OS, metadata.os);
                 deepStrictEqual(npmPackage.CPU, metadata.cpu);
+                deepStrictEqual(npmPackage.Exports, metadata.exports);
+                deepStrictEqual(npmPackage.Imports, metadata.imports);
                 strictEqual(npmPackage.Main, metadata.main);
                 strictEqual(npmPackage.Types, metadata.types);
                 deepStrictEqual(npmPackage.Browser, metadata.browser);
@@ -413,6 +433,7 @@ export function PackageTests(context: TestContext): void
                                     let assertions = [
                                         [nameof<Package>((pkg) => pkg.Name), null],
                                         [nameof<Package>((pkg) => pkg.Version), null],
+                                        [nameof<Package>((pkg) => pkg.Type), null],
                                         [nameof<Package>((pkg) => pkg.Private), null],
                                         [nameof<Package>((pkg) => pkg.Description), null],
                                         [nameof<Package>((pkg) => pkg.Author), {} as PackagePerson, AssertPerson],