+++ CHANGELOG.md
@@ -6,7 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## NPMPackageEditor [Unreleased]
 
-[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v3.0.0...dev)
+[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v3.0.1...dev)
+
+## NPMPackageEditor v3.0.1
+### Added
+  - Support for new `package.json` fields:
+    - `type` for specifying whether the package is a CommonJS or an ES Module
+    - `exports` for specifying entry points
+    - `imports` for declaring subpath imports
+
+### Updated
+  - All dependencies
+
+[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v3.0.0...v3.0.1)
 
 ## NPMPackageEditor v3.0.0
 ### Breaking